Split NewtonsoftJson as addon package
