+++ ObjectConfigurationExtensions.sln
@@ -11,6 +11,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{7D89563A
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests.Objects", "tests\Tests.Objects\Tests.Objects.csproj", "{06A9F0F7-10D7-4531-A346-2A18C5402C54}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Objects.NewtonsoftJson", "src\Objects.NewtonsoftJson\Objects.NewtonsoftJson.csproj", "{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -48,9 +50,22 @@ Global
 		{06A9F0F7-10D7-4531-A346-2A18C5402C54}.Release|x64.Build.0 = Release|Any CPU
 		{06A9F0F7-10D7-4531-A346-2A18C5402C54}.Release|x86.ActiveCfg = Release|Any CPU
 		{06A9F0F7-10D7-4531-A346-2A18C5402C54}.Release|x86.Build.0 = Release|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Debug|x64.Build.0 = Debug|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Debug|x86.Build.0 = Debug|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Release|x64.ActiveCfg = Release|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Release|x64.Build.0 = Release|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Release|x86.ActiveCfg = Release|Any CPU
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{A5A06279-A908-4B4D-A91B-919BDD4E2957} = {49A4B4B5-2EF8-422F-BC13-1622A3131D85}
 		{06A9F0F7-10D7-4531-A346-2A18C5402C54} = {7D89563A-B6B2-46E0-A640-E34CC6A2078A}
+		{9D57A59B-F955-4E36-ACEA-37C0D86C4A39} = {49A4B4B5-2EF8-422F-BC13-1622A3131D85}
 	EndGlobalSection
 EndGlobal
