+++ src/Objects.NewtonsoftJson/Objects.NewtonsoftJson.csproj
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <PropertyGroup>
-    <Authors>Renato Golia</Authors>    
+    <Authors>Renato Golia</Authors>
   </PropertyGroup>
 
   <PropertyGroup>
@@ -18,7 +18,7 @@
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <PackageProjectUrl>https://github.com/Kralizek/ObjectConfigurationExtensions</PackageProjectUrl>
     <PackageTags>dotnet-standard;microsoft;extensions;configuration;objects;</PackageTags>
-    <Description>Provider implementation for Microsoft.Extensions.Configuration to support concrete objects.</Description>
+    <Description>Extension of Kralizek.Extensions.Configuration using Newtonsoft.Json.</Description>
   </PropertyGroup>
 
   <PropertyGroup>
@@ -34,12 +34,10 @@
   </ItemGroup>
 
   <ItemGroup>
-    
+    <ProjectReference Include="..\Objects\Objects.csproj" />
   </ItemGroup>
-
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
-    <PackageReference Include="System.Text.Json" Version="8.0.0" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 
 </Project>
