+++ src/Objects/Internal/IConfigurationSerializer.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections.Generic;
+
+namespace Kralizek.Extensions.Configuration.Internal;
+
+public interface IConfigurationSerializer
+{
+    IDictionary<string, string?> Serialize(object source, string rootSectionName);
+}
