+++ tests/Tests.Objects/Internal/ObjectConfigurationSourceTests.cs
@@ -3,23 +3,22 @@ using Kralizek.Extensions.Configuration.Internal;
 using Microsoft.Extensions.Configuration;
 using NUnit.Framework;
 
-namespace Tests.Internal
+namespace Tests.Internal;
+
+[TestFixture]
+public class ObjectConfigurationSourceTests
 {
-    [TestFixture]
-    public class ObjectConfigurationSourceTests
+    [Test, CustomAutoData]
+    public void Constructor_is_guarded(GuardClauseAssertion assertion)
     {
-        [Test, CustomAutoData]
-        public void Constructor_is_guarded(GuardClauseAssertion assertion)
-        {
             assertion.Verify(typeof(ObjectConfigurationSource).GetConstructors());
         }
 
-        [Test, CustomAutoData]
-        public void Build_creates_a_provider(ObjectConfigurationSource sut, IConfigurationBuilder configurationBuilder)
-        {
+    [Test, CustomAutoData]
+    public void Build_creates_a_provider(ObjectConfigurationSource sut, IConfigurationBuilder configurationBuilder)
+    {
             var provider = sut.Build(configurationBuilder);
 
             Assert.That(provider, Is.InstanceOf<ObjectConfigurationProvider>());
         }
-    }
 }
