+++ tests/Tests.Objects/Tests.Objects.csproj
@@ -19,6 +19,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\src\Objects.NewtonsoftJson\Objects.NewtonsoftJson.csproj" />
     <ProjectReference Include="..\..\src\Objects\Objects.csproj" />
   </ItemGroup>
